pyproject.toml
[tool.poetry]
name = "repeng"
version = "0.1.0"
description = ""
authors = ["Misha Wagner <mishajw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
# Core ML/Data Science
torch = "^2.1.2"
transformers = "^4.36.2"
datasets = "^2.16.1"
accelerate = "^0.26.1"
scikit-learn = "^1.3.2"
# Data Processing
numpy = "^1.24.0"
pandas = "^2.0.0"
scipy = "^1.10.0"
# Plotting
matplotlib = "^3.7.0"
seaborn = "^0.13.1"
plotly = "^5.18.0"
kaleido = "0.2.1"
# Pipeline & Caching
mppr = { git = "https://github.com/mishajw/mppr.git", branch = "main" }
# AWS/S3
boto3 = "^1.28.0"
# Utilities
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
jaxtyping = "^0.2.25"
jsonlines = "^4.0.0"
fire = "^0.5.0"
tqdm = "^4.65.0"
# HuggingFace Optimization
hf_transfer = "^0.1.0"
# Notebooks
ipywidgets = "^8.0.0"
# Optional
openai = "^1.7.2"
pytest = "^7.4.4"
promptsource = { git = "https://github.com/bigscience-workshop/promptsource", branch = "main" }

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"
black = "^21.0.0"
pre-commit = "^3.6.0"
jupyter = "^1.0.0"
notebook = "^7.0.0"
jupyterlab = "^4.0.0"
pyright = "^1.1.347"
viztracer = "^0.16.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.ruff]
exclude = ['experiments/scratch']

[tool.pyright]
exclude = ['experiments/scratch']
